Fix cast exception

+++ toy-common/src/main/java/org/apache/toy/AbstractHBaseToy.java
@@ -21,8 +21,6 @@ import org.apache.toy.common.Parameter;
 
 import java.util.List;
 
-import static org.apache.toy.common.Parameter.UNSET;
-
 /**
  * HBase toy's base implementation. HBase configuration is inititlized in this class.
  */
@@ -35,7 +33,7 @@ public abstract class AbstractHBaseToy extends AbstractToy<Configuration> {
            if (parameter.type().equals(String.class))   parameter.checkAndSet(configuration.get(parameter.key()));
       else if (parameter.type().equals(String[].class)) parameter.checkAndSet(configuration.getStrings(parameter.key()));
       else if (parameter.type().isEnum())               parameter.checkAndSet(configuration.getEnum(parameter.key(), (Enum)parameter.value()));
-      else if (parameter.type().equals(Integer.class))  parameter.checkAndSet(configuration.getInt(parameter.key(), parameter.value()));
+      else if (parameter.type().equals(Integer.class))  parameter.checkAndSet(configuration.getInt(parameter.key(), (Integer)parameter.value()));
       else if (parameter.type().equals(Boolean.class))  parameter.checkAndSet(configuration.getBoolean(parameter.key(), (Boolean)parameter.value()));
 
       if (parameter.required() && parameter.empty()) {
